moved SessionStates to a session module
