+++ frost-client/src/lib.rs
@@ -5,4 +5,5 @@ pub mod client;
 pub mod coordinator;
 pub mod dkg;
 pub mod participant;
+pub mod session;
 pub mod trusted_dealer;
